requirements.txt
streamlit
groq
speechrecognition
gtts
python-dotenv
